requirements.txt
pandas
openpyxl
transformers
faiss-cpu
accelerate
sentencepiece
bitsandbytes